% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{boundbox}
\alias{boundbox}
\title{Formats boundbox for WFS query}
\usage{
boundbox(xmin, ymin, xmax, ymax, st_bbox = NULL)
}
\arguments{
\item{xmin}{numeric; minimum x coordinate (longitude)}

\item{ymin}{numeric; minimum y coordinate (latitude)}

\item{xmax}{numeric; maximum x coordinate (longitude)}

\item{ymax}{numeric; maximum y coordinate (latitude)}

\item{st_bbox}{bbox; object generated using \link[sf]{st_bbox}}
}
\value{
character string formatted for WFS query and use in \link[VicmapR]{read_layer_sf}
}
\description{
Convenience function that formats a spatial bounding box into a character string that can be passed to the WFS query.
The function accepts either a \code{bbox} from \link[sf]{st_bbox} or the coordinates for the box edges (\code{xmin}, \code{ymin}, \code{xmax}, \code{ymax}).
}
\examples{
#### Using an sf object ####
sf_data <- sf::st_read(system.file("shape/nc.shp", package = "sf"), quiet = TRUE)
boundbox(st_bbox = sf::st_bbox(sf_data))

#### Using coordinates ####
boundbox(xmin = 144.05, ymin = -38.44, xmax = 144.50, ymax = -38.10)
}
