% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_geoserver.R
\name{check_geoserver}
\alias{check_geoserver}
\title{Check Geoserver Response}
\usage{
check_geoserver(timeout = 15, quiet = FALSE)
}
\arguments{
\item{timeout}{numeric: the time (in seconds) to wait for the response before timing out (default is 15)}

\item{quiet}{logical: whether to silently check the connection and if working, return nothing. If \code{FALSE} (default),
the status message will be printed (\link[httr]{http_status})}
}
\value{
logical, TRUE if the geoserver is working
}
\description{
VicmapR relies upon a functioning geoserver. If for whatever reason the geoserver is not functioning then the functions
in this package will not work. This function will check the response of the geoserver; erroring out if the connection is down.
}
\examples{
\donttest{
check_geoserver()
}
}
