% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-query.R
\name{collect.vicmap_promise}
\alias{collect.vicmap_promise}
\title{collect}
\usage{
\method{collect}{vicmap_promise}(x, quiet = FALSE, paginate = TRUE, ...)
}
\arguments{
\item{x}{object of class \code{vicmap_promise} (likely passed from \code{\link[=vicmap_query]{vicmap_query()}})}

\item{quiet}{logical; whether to suppress the printing of messages and progress}

\item{paginate}{logical; whether to allow pagination of results to extract all records (default is TRUE)}

\item{...}{}
}
\value{

}
\description{
collect
}
\examples{
vicmap_query(layer = "datavic:VMHYDRO_WATERCOURSE_DRAIN") \%>\%
head(50) \%>\%
collect()
}
