% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-layer-names.R
\name{convert_layer_name}
\alias{convert_layer_name}
\title{Convert layer names from old platform naming convention to the new platform naming convention}
\usage{
convert_layer_name(x)
}
\arguments{
\item{x}{object of class \code{vicmap_promise}}
}
\value{
object of class \code{vicmap_promise}
}
\description{
Convert layer names from old platform naming convention to the new platform naming convention
}
\examples{
\donttest{
try(
vicmap_query(layer = "datavic:VMHYDRO_WATERCOURSE_DRAIN") 
)
}
}
