% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cql-predicates.R
\name{CQL}
\alias{CQL}
\title{CQL escaping}
\usage{
CQL(...)
}
\arguments{
\item{...}{Character vectors that will be combined into a single CQL statement.}
}
\value{
An object of class \code{c("CQL", "SQL")}
}
\description{
Write a CQL expression to escape its inputs, and return a CQL/SQL object.
Used when writing filter expressions in \code{\link[=vicmap_query]{vicmap_query()}}.
}
\details{
See \href{https://docs.geoserver.org/stable/en/user/tutorials/cql/cql_tutorial.html}{the CQL/ECQL for Geoserver website}.

The code for cql escaping was developed by the bcdata team: \url{https://bcgov.github.io/bcdata/reference/cql_geom_predicates.html}
}
\examples{
CQL("FOO > 12 & NAME LIKE 'A&'")
}
