% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listFeatures.R
\name{extract_abstract}
\alias{extract_abstract}
\title{get abstract df from content of API response}
\usage{
extract_abstract(c)
}
\arguments{
\item{c}{content from metashare api}
}
\description{
get abstract df from content of API response
}
