% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-metadata.R
\name{feature_hits}
\alias{feature_hits}
\title{number of rows}
\usage{
feature_hits(x)
}
\arguments{
\item{x}{object of class \code{vicmap_promise}}
}
\value{
numeric
}
\description{
number of rows
}
\examples{
vicmap_query(layer = "datavic:VMHYDRO_WATERCOURSE_DRAIN") \%>\%
 feature_hits()
}
