% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillter.R
\name{filter.vicmap_promise}
\alias{filter.vicmap_promise}
\title{filter}
\usage{
\method{filter}{vicmap_promise}(x, ...)
}
\arguments{
\item{x}{object of class \code{vicmap_promise} (likely passed from \code{\link[=vicmap_query]{vicmap_query()}})}

\item{...}{filter statements}
}
\value{

}
\description{
filter
}
