% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-metadata.R
\name{feature_cols}
\alias{feature_cols}
\alias{geom_col_name}
\title{feature column names}
\usage{
feature_cols(x)

geom_col_name(x)
}
\arguments{
\item{x}{object of class \code{vicmap_promise}}
}
\value{
character

character
}
\description{
feature column names

geom column name
}
\section{Functions}{
\itemize{
\item \code{feature_cols}: 
}}

\examples{
vicmap_query(layer = "datavic:VMHYDRO_WATERCOURSE_DRAIN") \%>\% feature_cols()
vicmap_query(layer = "datavic:VMHYDRO_WATERCOURSE_DRAIN") \%>\% 
geom_col_name()
}
