% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-metadata.R
\name{get_col_df}
\alias{get_col_df}
\title{return df of column names and types}
\usage{
get_col_df(x)
}
\arguments{
\item{x}{object of class \code{vicmap_promise}}
}
\value{
data.frame
}
\description{
return df of column names and types
}
\examples{
vicmap_query(layer = "datavic:VMHYDRO_WATERCOURSE_DRAIN") \%>\% 
get_col_df()
}
