% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-query.R
\name{head.vicmap_promise}
\alias{head.vicmap_promise}
\title{Return the first n rows of the data}
\usage{
\method{head}{vicmap_promise}(x, n = 5, ...)
}
\arguments{
\item{x}{object of class \code{vicmap_promise} (likely passed from \code{\link[=vicmap_query]{vicmap_query()}})}

\item{n}{integer; number of rows to return}

\item{...}{Other parameters possibly used by generic}
}
\value{
vicmap_promise
}
\description{
Return the first n rows of the data
}
\examples{
\dontrun{
vicmap_query(layer = "datavic:VMHYDRO_WATERCOURSE_DRAIN") \%>\%
head(50)
}
}
