% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{newClient}
\alias{newClient}
\title{New WFSClient}
\usage{
newClient(version = "1.1.1", ...)
}
\arguments{
\item{version}{Version of WFS to use}

\item{...}{Additional arguments passed to \link[ows4R]{WFSClient}$new}
}
\value{
Object of \link[R6]{R6Class} with methods for interfacing an OGC Web Feature Service.
}
\description{
Establishes a new WFSClient for Victoria's spatial data using \link[ows4R]{WFSClient}
}
\examples{
VicmapClient <- newClient()
}
