% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygonFormat.R
\name{polygonFormat.sf}
\alias{polygonFormat.sf}
\title{Simplify polygon}
\usage{
\method{polygonFormat}{sf}(shape, ...)
}
\arguments{
\item{shape}{}

\item{...}{}
}
\value{

}
\description{
Simplify polygon
}
