% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-query.R
\name{print.vicmap_promise}
\alias{print.vicmap_promise}
\title{print}
\usage{
\method{print}{vicmap_promise}(x, ...)
}
\arguments{
\item{x}{object of class \code{vicmap_promise} (likely passed from \code{\link[=vicmap_query]{vicmap_query()}})}

\item{...}{arguments to be passed to \link[base]{print}}
}
\value{
vicmap_promise (invisible), promise sample printed to console
}
\description{
print
}
\examples{
query <- vicmap_query(layer = "datavic:VMHYDRO_WATERCOURSE_DRAIN")
print(query)
}
