% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWFS.R
\name{read_layer_sf}
\alias{read_layer_sf}
\title{Read Vicmap WFS layer as sf}
\usage{
read_layer_sf(
  layer_name,
  filter = NULL,
  boundbox = NULL,
  CRS = 4283,
  Client = newClient(),
  ...
)
}
\arguments{
\item{layer_name}{character; name of Vicmap layer/feature to read in.}

\item{filter}{character; \strong{optional} CQL filter statement. See \link[https://docs.geoserver.org/stable/en/user/tutorials/cql/cql_tutorial.html]{CQL tutorial} for more information.}

\item{boundbox}{character; \strong{optional} string of coordinates to restrict query within. Can be generated using the \link[VicmapR]{boundbox} function or formatted as a character string following: \code{"xmin,ymin,xmax,ymax"}.}

\item{CRS}{numeric; coordinate reference system for query (default is 4283).}

\item{Client}{Object of \link[R6]{R6Class}; Vicmap client generated with \link[VicmapR]{newClient}. Mandatory when using a boundbox; however default is \link[VicmapR]{newClient}}

\item{...}{Additional arguments passed to \link[sf]{read_sf} (e.g. \code{quiet} and \code{as_tibble}).}
}
\value{
sf/data.frame/tibble
}
\description{
Reads in a Vicmap layer as an \link[sf]{sf}. Filters can be made using a bounding box or with additional CQL filtering statements.
}
\examples{
VicmapClient <- newClient()
data <- read_layer_sf(layer_name = "datavic:VMHYDRO_WATERCOURSE_DRAIN", 
                      boundbox = boundbox(xmin = 144.25, 
                                          ymin = -38.44, 
                                          xmax = 144.50, 
                                          ymax = -38.25), 
                      filter = "HIERARCHY = 'L'", 
                      Client = VicmapClient)
}
