% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-route.R
\docType{data}
\name{route}
\alias{route}
\title{Cycling Route Data}
\format{
An object of class \code{data.frame} with 3137 rows and 11 columns.
}
\source{
\href{https://ridewithgps.com/routes/4120656}{ridewithGPS}
}
\usage{
data(route)
}
\description{
A dataframe containing coordinates long a cycling route. Used as example data in the introduction to Vicmap vignette.
}
\examples{
data(route)
}
\keyword{datasets}
