% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select.vicmap_promise}
\alias{select.vicmap_promise}
\title{select}
\usage{
\method{select}{vicmap_promise}(x, ...)
}
\arguments{
\item{x}{object of class \code{vicmap_promise} (likely passed from \code{\link[=vicmap_query]{vicmap_query()}})}

\item{...}{}
}
\value{

}
\description{
select
}
