% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-metadata.R
\name{specify_geom_name}
\alias{specify_geom_name}
\title{apply cql to geom}
\usage{
specify_geom_name(x, CQL_statement)
}
\arguments{
\item{x}{}

\item{CQL_statement}{}
}
\value{

}
\description{
apply cql to geom
}
