% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-query.R
\name{vicmap_query}
\alias{vicmap_query}
\title{vicmap_query}
\usage{
vicmap_query(layer, CRS = 4283)
}
\arguments{
\item{layer}{vicmap layer to query. Options are listed in `VicmapR::listLayers()``}

\item{CRS}{Coordinate Reference System (default is 4283)}
}
\value{

}
\description{
vicmap_query
}
\examples{
vicmap_query(layer = "datavic:VMHYDRO_WATERCOURSE_DRAIN")
}
