% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-metadata.R
\name{feature_hits}
\alias{feature_hits}
\title{The Number of Rows of the Promised Data}
\usage{
feature_hits(x)
}
\arguments{
\item{x}{object of class \code{vicmap_promise}}
}
\value{
integer
}
\description{
\code{feature_hits()} returns an integer of the number of rows that match the passed query/promise.
This is similar to how \code{nrow()} works for a data.frame, however it will evaluate the number of rows to be returned
without having to download the data.
}
\examples{
\donttest{
vicmap_query(layer = "datavic:VMHYDRO_WATERCOURSE_DRAIN") \%>\%
 feature_hits()
 }
}
