% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter.vicmap_promise}
\alias{filter.vicmap_promise}
\title{filter}
\usage{
\method{filter}{vicmap_promise}(.data, ...)
}
\arguments{
\item{.data}{object of class \code{vicmap_promise} (likely passed from \code{\link[=vicmap_query]{vicmap_query()}})}

\item{...}{filter statements}
}
\value{
object of class \code{vicmap_promise}
}
\description{
filter
}
\examples{
vicmap_query(layer = "datavic:VMHYDRO_WATERCOURSE_DRAIN") \%>\%
 filter(HIERARCHY == "L", PFI == 8553127)
}
