% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-query.R, R/utils-head.R
\name{head.vicmap_promise}
\alias{head.vicmap_promise}
\alias{head}
\title{Return the first n rows of the data}
\usage{
\method{head}{vicmap_promise}(x, n = 5, ...)
}
\arguments{
\item{x}{object of class \code{vicmap_promise} (likely passed from \code{\link[=vicmap_query]{vicmap_query()}})}

\item{n}{integer; number of rows to return}

\item{...}{Other parameters possibly used by generic}
}
\value{
Object of class \code{vicmap_promise}, which is a 'promise' of the data that can  be returned if \code{collect()} is used
}
\description{
See \code{utils::\link[utils]{head}} for details.
}
\section{Methods (by class)}{
\itemize{
\item \code{vicmap_promise}: head.vicmap_promise
}}

\examples{
\donttest{
try(
vicmap_query(layer = "open-data-platform:hy_watercourse") \%>\%
head(50)
)
}
}
\keyword{internal}
