% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{listLayers}
\alias{listLayers}
\alias{listFields}
\title{List available WFS layers}
\usage{
listLayers(Client, ...)

listFields(Client, layer_name)
}
\arguments{
\item{Client}{WFSClient generated by \link[VicmapR]{newClient}}

\item{...}{Additional arguments passed to \link[stringr]{str_detect}. The \code{pattern} argument can be used to search for specific layers with matching names or titles.}

\item{layer_name}{character; name of the WFS layer}
}
\value{
data.frame

character; fields/columns of data
}
\description{
Wraps \code{getFeatureTypes} from \link[ows4R]{WFSClient} and allows for string searching of layers
}
\examples{
VicmapClient <- newClient()
listLayers(VicmapClient, pattern = stringr::regex("flood height contour", ignore_case = TRUE))
VicmapClient <- newClient()
listFields(VicmapClient, "datavic:VMLITE_TR_AIRPORT")
}
