% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listFeatures.R
\name{listLayers}
\alias{listLayers}
\title{List available WFS layers}
\usage{
listLayers(...)
}
\arguments{
\item{...}{Additional arguments passed to \link[stringr]{str_detect}. The \code{pattern} argument can be used to search for specific layers with matching names or titles.}
}
\value{
data.frame
}
\description{
Lists layers available from Vicmap
}
\examples{
listLayers(pattern = stringr::regex("flood height contour", ignore_case = TRUE))
}
