% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-query.R
\name{print.vicmap_promise}
\alias{print.vicmap_promise}
\title{Print a Snapshot of the Data}
\usage{
\method{print}{vicmap_promise}(x, ...)
}
\arguments{
\item{x}{object of class \code{vicmap_promise} (likely passed from \code{\link[=vicmap_query]{vicmap_query()}})}

\item{...}{arguments to be passed to \link[base]{print}}
}
\value{
vicmap_promise (invisible), promise sample printed to console
}
\description{
\code{print()} displays a cut of the data (no more than  six rows)
alongside the number of rows and columns that would be returned.
}
\examples{
\donttest{
try(
query <- vicmap_query(layer = "datavic:VMHYDRO_WATERCOURSE_DRAIN")
)
try(
print(query)
)
}
}
