% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-query.R
\name{show_query.vicmap_promise}
\alias{show_query.vicmap_promise}
\title{show_query}
\usage{
\method{show_query}{vicmap_promise}(x, ...)
}
\arguments{
\item{x}{object of class \code{vicmap_promise} (likely passed from \code{\link[=vicmap_query]{vicmap_query()}})}
}
\value{
vicmap_promise (invisible), query printed to console
}
\description{
show_query
}
\examples{
vicmap_query(layer = "datavic:VMHYDRO_WATERCOURSE_DRAIN") \%>\%
head(50) \%>\%
show_query()
}
